Move ensureIsArray form parser utils to a array utils
