+++ src/component/registry/bpmn-elements-registry.ts
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { ensureIsArray } from '../parser/json/converter/utils';
+import { ensureIsArray } from '../helpers/array-utils';
 import { BpmnMxGraph } from '../mxgraph/BpmnMxGraph';
 import { computeBpmnBaseClassName, extractBpmnKindFromStyle } from '../mxgraph/style-helper';
 import { FlowKind } from '../../model/bpmn/internal/edge/FlowKind';
